# openapi.yaml
openapi: 3.0.3
info:
  title: Whop SaaS Marketplace API
  description: |
    Production-grade marketplace API enabling multi-tenant SaaS, Stripe Connect payments,
    license management, and webhook event broadcasting.

    ## Authentication
    All endpoints require authentication via session token cookie (`session_token`).

    ## Multi-Tenancy
    All resources are scoped to companies. Users can access multiple companies with different roles.

    ## Rate Limiting
    - 100 requests per minute per IP
    - Returns `429 Too Many Requests` when exceeded

    ## Security
    - Tenant isolation via `company_id` filtering
    - Role-based access control (owner/admin/member)
    - Money stored in minor units (cents) to avoid floating-point errors
  version: 1.0.0
  contact:
    name: API Support
    email: api@whopsaas.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.whopsaas.com
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: JWT session token (HTTP-only cookie)
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod_abc123
        company_id:
          type: string
          example: biz_xyz789
        title:
          type: string
          example: Pro Plan
        description:
          type: string
          nullable: true
          example: Premium features with priority support
        image_url:
          type: string
          format: uri
          nullable: true
        price_amount:
          type: number
          format: decimal
          example: 49.99
        price_minor_units:
          type: integer
          example: 4999
          description: Price in cents (avoid floating-point errors)
        currency:
          type: string
          example: USD
        plan_type:
          type: string
          enum: [one_time, monthly, yearly]
        billing_period:
          type: integer
          nullable: true
          description: Billing period in days (30, 365, etc.)
        trial_days:
          type: integer
          example: 14
        active:
          type: boolean
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - company_id
        - title
        - price_amount
        - price_minor_units
        - currency
        - plan_type
...
    Company:
      type: object
      properties:
        id:
          type: string
          example: biz_abc123
        type:
          type: string
          enum: [platform, sub_merchant]
        status:
          type: string
          enum: [active, pending_kyc, suspended, closed]
        title:
          type: string
        email:
          type: string
          format: email
        slug:
          type: string
          example: my-saas-company
        logo_url:
          type: string
          format: uri
          nullable: true
        website_url:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        platform_fee_percent:
          type: number
          format: decimal
          example: 5.0
        payout_minimum_amount:
          type: number
          format: decimal
          example: 50.0
        payout_frequency:
          type: string
          enum: [daily, weekly, monthly]
        stripe_account_id:
          type: string
          nullable: true
        stripe_onboarded:
          type: boolean
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            products:
              type: integer
            memberships:
              type: integer
            payments:
              type: integer
            users:
              type: integer
    CompanyCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        type:
          type: string
          enum: [platform, sub_merchant]
          default: sub_merchant
        logo_url:
          type: string
          format: uri
        website_url:
          type: string
          format: uri
        description:
          type: string
        platform_fee_percent:
          type: number
          default: 5.0
        payout_minimum_amount:
          type: number
          default: 50.0
        payout_frequency:
          type: string
          enum: [daily, weekly, monthly]
          default: weekly
        metadata:
          type: object
      required:
        - title
        - email
    CompanyUpdate:
      type: object
      properties:
        title:
          type: string
        email:
          type: string
          format: email
        logo_url:
          type: string
          format: uri
          nullable: true
        website_url:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        platform_fee_percent:
          type: number
        payout_minimum_amount:
          type: number
        payout_frequency:
          type: string
          enum: [daily, weekly, monthly]
        status:
          type: string
          enum: [active, pending_kyc, suspended, closed]
          description: Only owners can change status
        metadata:
          type: object
          nullable: true

    # ========================================================================
    # Membership Schemas
    # ========================================================================
    Membership:
      type: object
      properties:
        id:
          type: string
          example: mem_abc123
        user_id:
          type: string
        product_id:
          type: string
        company_id:
          type: string
        status:
          type: string
          enum: [active, past_due, canceled, expired, trialing]
        stripe_subscription_id:
          type: string
          nullable: true
        current_period_start:
          type: string
          format: date-time
          nullable: true
        current_period_end:
          type: string
          format: date-time
          nullable: true
        cancel_at:
          type: string
          format: date-time
          nullable: true
        canceled_at:
          type: string
          format: date-time
          nullable: true
        trial_end:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
            avatar_url:
              type: string
        product:
          $ref: '#/components/schemas/Product'
        company:
          $ref: '#/components/schemas/Company'
    MembershipUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, past_due, canceled, expired, trialing]
          description: Only company admins can change status
        cancel_at:
          type: string
          format: date-time
          nullable: true
          description: Schedule cancellation (members can cancel own)
        metadata:
          type: object
          nullable: true
          description: Only company admins can update metadata

    # ========================================================================
    # Webhook Schemas
    # ========================================================================
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
security:
  - cookieAuth: []
paths:
...
  /api/companies:
    get:
      tags:
        - Companies
      summary: List companies
      description: List all companies user has access to
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [owner, admin, member]
        - name: type
          in: query
          schema:
            type: string
            enum: [platform, sub_merchant]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  companies:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Company'
                        - type: object
                          properties:
                            user_role:
                              type: string
                              enum: [owner, admin, member]
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Companies
      summary: Create company
      description: Create a new company (user becomes owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreate'
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '409':
          description: Conflict (email already exists)
  /api/companies/{id}:
    get:
      tags:
        - Companies
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Company'
                  - type: object
                    properties:
                      user_role:
                        type: string
        '404':
          description: Company not found

    put:
      tags:
        - Companies
      summary: Update company
      description: Update company (requires admin/owner role)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUpdate'
      responses:
        '200':
          description: Company updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '403':
          description: Forbidden
        '409':
          description: Conflict (email already exists)

  # ==========================================================================
  # MEMBERSHIPS
  # ==========================================================================
...
  /api/memberships/{id}:
    get:
      tags:
        - Memberships
      summary: Get membership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '404':
          description: Membership not found

    put:
      tags:
        - Memberships
      summary: Update membership
      description: Update membership status or metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipUpdate'
      responses:
        '200':
          description: Membership updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '403':
          description: Forbidden

  # ==========================================================================
  # WEBHOOKS
  # ==========================================================================
...